requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.7            # via django
django-background-tasks==1.2.5  # via -r requirements.in
django-bootstrap3==14.0.0  # via -r requirements.in
django-compat==1.0.15     # via django-background-tasks
django==3.0.3             # via -r requirements.in, django-bootstrap3, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
pillow==7.1.2             # via -r requirements.in
pytz==2020.1              # via django
six==1.15.0               # via django-background-tasks, django-compat
sqlparse==0.3.1           # via django
